cleanup unused functions and subcommands
