+++ pyproject.toml
@@ -29,5 +29,6 @@ classifiers = [
 "Homepage" = "https://github.com/asfadmin/AI-Event-Monitoring/tree/main"
 [project.scripts]
 insar-eventnet = "insar_eventnet.aievents:cli"
+sagemaker = "insar_eventnet.sagemaker:sagemaker_server"
 [tool.setuptools]
 packages = ["insar_eventnet", "insar_eventnet.architectures"]
